requirements.txt
# --- Build tools ---
setuptools>=75.0.0

# --- Core analytics (kept from project) ---
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# --- Jupyter ---
jupyter>=1.0.0,<2.0.0
ipykernel>=6.0.0,<7.0.0

#XGBoost Model
xgboost>=2.0.0,<3.0.0

# Model explainability
shap>=0.45.0

# Pin TF/Keras pair to keep `from keras import ...` API working
tensorflow==2.15.*
keras==2.15.*

# (No need to list: pathlib, ast — they are in the Python stdlib)
pyyaml>=6.0,<7.0
